Added Submission Files & Sample predict.py
All the files in the Submission Folder are the ones the TA sent.
The predict.py is where we will send the test data to the trained model to produce a decision/prediction whether the handwriting is that of a male or female. Also, measures the time taken to produce that decision.
predict.py was tested and it's running. (Doesn't cause runtime or compilation errors). It doesn't create any folders if they don't exist though.